requirements.txt
pillow
matplotlib
pandas
numpy
torch
torchvision
scikit-learn
jupyter